consolidating testing code
